Update Client_GUI.py
Added code for additional robustness. Code added includes a close client method. Whenever an error occurs, the program safely closes and resets the GUI so the user can try again. Also, added robustness to the port number. If digits are entered, it checks if there's too many digits and checks if any characters are entered that may cause a datatype error.